Add files via upload
Op Sys Memory Manager, used linked lists and singletons to manage the memory creation, deletion, usage, and freeing. Used junit and native calls to convert c++ methods and functions into junit methods as the main language in the GUI set on Android Studio. Tested many cases and passed the majority of the cases.